# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PetChain API",
    "version": "1.0.0",
...
    "contact": {
      "name": "PetChain Engineering",
      "email": "engineering@petchain.app",
      "url": "https://petchain.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    },
    {
      "url": "https://staging.petchain.app/api",
      "description": "Staging environment"
    },
    {
      "url": "https://api.petchain.app/api",
      "description": "Production environment"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users",
        "description": "Returns a paginated list of users. Requires admin role.",
        "operationId": "listUsers",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "$ref": "#/components/parameters/SortByParam"
          },
          {
            "$ref": "#/components/parameters/SortOrderParam"
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["owner", "vet", "admin"]
            },
            "description": "Filter by user role"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search by name or email (case-insensitive)",
            "example": "jane"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/pets": {
      "get": {
        "tags": ["Pets"],
        "summary": "List pets",
        "description": "Returns a paginated list of pets. Owners see only their own pets; vets and admins can see all pets.",
        "operationId": "listPets",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "$ref": "#/components/parameters/SortByParam"
          },
          {
            "$ref": "#/components/parameters/SortOrderParam"
          },
          {
            "name": "ownerId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter pets by owner ID",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          {
            "name": "species",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["dog", "cat", "bird", "rabbit", "other"]
            },
            "description": "Filter by species"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search by pet name (case-insensitive)",
            "example": "buddy"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of pets",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Pet"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Pets"],
        "summary": "Create a new pet",
        "description": "Registers a new pet. The `ownerId` must match the authenticated user's ID (or the user must be an admin).",
        "operationId": "createPet",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetRequest"
              },
              "example": {
                "name": "Buddy",
                "species": "dog",
                "breed": "Golden Retriever",
                "dateOfBirth": "2021-06-15",
                "microchipId": "985141002512345",
                "ownerId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pet created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Pet"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/medical-records": {
      "get": {
        "tags": ["Medical Records"],
        "summary": "List medical records",
        "description": "Returns a paginated list of medical records with optional filters.",
        "operationId": "listMedicalRecords",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "$ref": "#/components/parameters/SortByParam"
          },
          {
            "$ref": "#/components/parameters/SortOrderParam"
          },
          {
            "name": "petId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by pet ID"
          },
          {
            "name": "vetId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by veterinarian ID"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["checkup", "vaccination", "surgery", "treatment", "other"]
            },
            "description": "Filter by record type"
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter records from this date (inclusive)",
            "example": "2024-01-01"
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter records up to this date (inclusive)",
            "example": "2024-12-31"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of medical records",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MedicalRecord"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Medical Records"],
        "summary": "Create a medical record",
        "description": "Creates a new medical record for a pet. After creation, the record hash is automatically anchored on the Stellar blockchain asynchronously. The `isBlockchainVerified` field will be `false` until anchoring completes.",
        "operationId": "createMedicalRecord",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMedicalRecordRequest"
              },
              "example": {
                "petId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                "vetId": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                "type": "vaccination",
                "diagnosis": "Healthy — routine annual checkup",
                "treatment": "Administered DHPP booster vaccine",
                "notes": "Pet is in excellent health. Weight stable at 28kg.",
                "visitDate": "2024-03-15",
                "nextVisitDate": "2025-03-15",
                "vaccinations": [
                  {
                    "vaccineName": "DHPP",
                    "administeredAt": "2024-03-15",
                    "nextDueDate": "2025-03-15",
                    "manufacturer": "Merck Animal Health",
                    "batchNumber": "LOT-2024-001",
                    "dose": "1ml"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Medical record created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MedicalRecord"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Pet or vet not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/appointments": {
      "get": {
        "tags": ["Appointments"],
        "summary": "List appointments",
        "description": "Returns appointments. Owners see their pets' appointments; vets see their own appointments; admins see all.",
        "operationId": "listAppointments",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "petId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by pet ID"
          },
          {
            "name": "vetId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by vet ID"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["PENDING", "CONFIRMED", "CANCELLED", "COMPLETED", "NO_SHOW", "RESCHEDULED"]
            },
            "description": "Filter by appointment status"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ROUTINE_CHECKUP",
                "VACCINATION",
                "SURGERY",
                "DENTAL",
                "GROOMING",
                "EMERGENCY",
                "FOLLOW_UP",
                "DIAGNOSTIC",
                "SPECIALIST_REFERRAL",
                "NUTRITION_CONSULTATION"
              ]
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter appointments from this date (inclusive)",
            "example": "2024-03-01"
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter appointments up to this date (inclusive)",
            "example": "2024-03-31"
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "List of appointments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Appointments"],
        "summary": "Create an appointment",
        "description": "Books a new appointment for a pet with a veterinarian.",
        "operationId": "createAppointment",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentRequest"
              },
              "example": {
                "petId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                "vetId": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                "date": "2024-03-15",
                "time": "14:30",
                "durationMinutes": 30,
                "type": "VACCINATION",
                "status": "PENDING",
                "notes": "Annual booster vaccines due",
                "reminder": {
                  "isEnabled": true,
                  "minutesBefore": 60,
                  "notificationMethod": "push"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Appointment"
                    },
                    "message": {
                      "type": "string",
                      "example": "Appointment booked successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Pet or vet not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Time slot already booked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "SLOT_UNAVAILABLE",
                    "message": "The requested time slot is already booked"
                  },
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/medications": {
      "get": {
        "tags": ["Medications"],
        "summary": "List medications",
        "description": "Returns medications for a pet. The `petId` query parameter is required.",
        "operationId": "listMedications",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "petId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Pet ID to fetch medications for",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "paused", "completed", "discontinued"]
            },
            "description": "Filter by medication status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of medications",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Medication"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing required petId parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Medications"],
        "summary": "Create a medication schedule",
        "description": "Creates a new medication schedule for a pet.",
        "operationId": "createMedication",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMedicationRequest"
              },
              "example": {
                "petId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                "name": "Apoquel",
                "dosage": "16mg",
                "frequency": "once_daily",
                "durationDays": 30,
                "startDate": "2024-03-15",
                "endDate": "2024-04-14",
                "status": "active",
                "instructions": "Give with food. Monitor for side effects."
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Medication schedule created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Medication"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Pet not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from `POST /auth/login` or `POST /auth/register`.\n\n**Example:**\n```\nAuthorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n```\n\nTokens expire after 7 days. Use `POST /auth/refresh` to obtain a new token."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for server-to-server integrations. Contact engineering@petchain.app to obtain a key."
      }
    },
    "schemas": {
...
      "PaginatedResponse": {
        "type": "object",
        "required": ["success", "data", "pagination", "timestamp"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {}
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "description": "Page number for pagination"
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        },
        "description": "Number of items per page"
      },
      "SortByParam": {
        "name": "sortBy",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Field to sort by (e.g. \"createdAt\", \"name\")"
      },
      "SortOrderParam": {
        "name": "sortOrder",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": ["asc", "desc"],
          "default": "desc"
        },
        "description": "Sort direction"
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized — missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required. Please log in."
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}